Fixed animation error, started Animation class
Turns out the problem with the textures was that the load identity was
in the wrong location... I started working on the Animation class so
that it can go to the next frame, the previous frame, set the frame to a
certain number, and get a certain frame from a number. I'm hooking up
the player class with animations to see how it goes.