Animation works, need way to control speed
I set up the Animation class and created a loader using folders as the
base. Basically you load all the images for an animation into a folder
and then load that folder. It then searches through the folder looking
for sprite_#.png folders, and it adds it to an ArrayList of textures to
be played back.